+++ CHANGES.md
@@ -14,6 +14,7 @@
     - `BatchYielder` now takes an `input_mask` argument to filter inputs
     - `Model` prediction methods no longer take input mask arguments, instead the input mask (if present) is automatically used. If users have already filtered their data, they should manually remove the input mask from the model (i.e. set it to None)
 - Callbacks which take arguments related to (sub-)epochs (e.g. cycle length, scale, time to renewal. etc. for `CycleLR`, `OneCycle`, etc. and `SWA`) now take these arguments in terms of epochs. I.e. a OneCycle schedule with 9 training folds, running for 15 epochs would previously require e.g. `lenghts=(45,90)` in order to complete the cycle in 15 epochs (135 subepochs). Now it is specified as simply `lenghts=(5,10)`. Additionally, these arguments must be integers. Floats will be coerced to integers with warning.
+- `lr_find` now runds over all training folds, instead of just 1
   
 ## Breaking
 
@@ -52,6 +53,7 @@
     - `fit` now expects to perform the entire fitting of the model, rather than just one sup-epoch. Additionally, validation loss is now computed only at the end of the epoch, rather that previously where it was computed after each fold.
 - `SWA` `renewal_period` should now be None in order to prevent a second average being tracked (previously was negative)
 - Some examples have been renamed, and copies using the old model fitting proceedure and old callbacks are available in `examples/old`
+- `lr_find` now runds over all training folds, instead of just 1
 
 ## Depreciations